==> R/GDSC/01_cpd_PCAfeatures.ipynb <==
# 0. Setup

# Verify we're in the correct working directory
import os
import git 
from pathlib import Path

def get_project_root():
    return Path(git.Repo('.', search_parent_directories=True).working_tree_dir)

root = get_project_root()

os.chdir(root)
os.getcwd()

plot_folder = "images/GDSC/"

## import packages, models, trainers

import argparse
import logging
import sys

import numpy as np
import pandas as pd


from scripts_model.cpd_smiles_embed import *

import random
seed=42

random.seed(seed)
np.random.seed(seed)
os.environ['PYTHONHASHSEED'] = str(seed)

# 1. Prepare dataset

## Load 

data_folder = "data/GDSC"

cpd_smiles = pd.read_csv(os.path.join(data_folder, "gdsc_cpd_metadata_w_smiles.csv"))

## Get PCA-based features

cpd_smiles = cpd_smiles[['drug_id', 'smiles']]
cpd_smiles = cpd_smiles.set_index('drug_id')

# get PCA-based features
d_data = smiles_to_AtonBondDescriptor_PCAembedings(cpd_smiles)
d_data.index = d_data.index.astype(str)

d_data.shape

d_data.to_csv(os.path.join(data_folder, "GDSC_cpd_pca_features.csv"), index=True)
